fixed error caused by no avatar

+++ server/routes/api.js
@@ -53,24 +53,31 @@ const isAuthenticated = (req, res, next) => {
 // Adds/changes user profile picture
 router.post("/addProfilePicture", isAuthenticated, upload.single("image"), async (req, res) => {
   try {
+    console.log(req.file.path)
     // Find user who made the request
     const user = await Users.findOne({email: res.locals.userEmail})
+    console.log(user)
 
     // Delete old picture file if exists
     if (user.avatarFile) {
-      fs.unlinkSync("./uploads/"+user.avatarFile)
+      const files = fs.readdirSync("./uploads/")
+      console.log(files)
+      if (files.includes(user.avatarFile)) {
+        fs.unlinkSync("./uploads/"+user.avatarFile)
+      }
     }
 
     // Update picture file name for user document
     if (req.file.path.split("\\")[1]) {
+      console.log(req.file.path.split("\\")[1])
       await Users.updateOne(
         {email: res.locals.userEmail}, 
         {$set: {avatarFile: req.file.path.split("\\")[1]}}
       )
     }
-    
     // Form full url for front-end and send it back
     const avatarPath = req.protocol+'://'+req.get('host')+"/"+req.file.path
+    console.log(avatarPath)
     res.send({
       success: true,
       path: avatarPath
